pyproject.toml
[project]
name = "juan-mas-app"
version = "0.1.0"
description = "Databricks Chatbot App - Works locally and on Databricks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash==3.0.2",
    "dash-bootstrap-components==2.0.0",
    "mlflow>=2.21.2",
    "python-dotenv==1.1.0",
    "databricks-sdk",
]


